Only attempt to execute mojo as part of download test if OS is linux.
This is because the download URL in test is hardcoded to linux.
The test fails in MacOS for example, if run.
To avoid this, assume that OS is Linux before execution.
Otherwise, do not execute the mojo.

A better approach would be to programmatically download the relevant
binary. I thought that should be done in a separate PR. For now the
simplest solution is to only run the mojo when it makes sense to.