Initial pass at supporting Helm plugins
Provides two new goals, install-plugin and invoke-plugin, the former
helps with installing plugins while the latter allows them to actually
be invoked as part of the build lifecycle.  This is intended to allow
for using plugins such as the awesome helm unittest plugin as part of
your builds.

This has some relevance to addressing #86